requirements.txt
pandas
python-dotenv
psycopg[binary]